Fix a race condition and potential orphaned nodes
Fixed a race condition that could lead to a crash when allocate_block() is called from allocate.  If a thread blocked on the spin lock in allocate_block() and returned because m_free.node was no longer null, it's possible that other threads could consume all of the memory that had just been allocated, leading to a dereference of nullptr in allocate()

Also, in allocate_block() it's possible that threads returning memory to the pool could do so while allocate block was allocating new memory.  in that case, the code would orphan all such nodes.